文章发布

+++ app/Http/Controllers/Admin/CateController.php
@@ -183,6 +183,13 @@ class CateController extends Controller
             ], 200);
         }
 
+        if ($cate->hasposts()) {
+            return response()->json([
+                'success' => false,
+                'message' => '请先删除分类下的文章',
+            ], 200);
+        }
+
         // 删除图片
         if ($cate->image) {
             Storage::disk('public')->delete($cate->image);
